requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
pydantic==2.5.0
pydantic-settings==2.1.0
fastapi-limiter==0.1.6
redis==5.0.1
python-multipart==0.0.6
starlette==0.27.0
